fix: when backfilling an index, don't delete other indexes

+++ share/search/index_strategy/trovesearch_denorm.py
@@ -1000,7 +1000,9 @@ def task__delete_iri_value_scraps(
         _index.full_index_name
         for _index in _index_strategy.each_live_index(any_strategy_check=True)
         if _index.subname == 'iri_values'
-    }
+    }.intersection(indexnames)
+    if not _irivalue_indexnames:
+        return
     # delete any docs that belong to cards in this chunk but weren't touched by indexing
     _delete_resp = _index_strategy.es8_client.delete_by_query(
         index=list(_irivalue_indexnames),
